Remove older workaround to view private and hidden posts in BuddyPress forums.
Previously, commits c598354 and 255d626 were needed to search for
posts in a private or hidden group forum. That is no longer necessary
in the latest version of bbPress.

The has_access() method is still kept around to avoid leaking the
private (or hidden) group forum name in the header from the main
bbPress search interface.